Move papernum detection to query prep
