+++ search/controllers/tests.py
@@ -73,16 +73,3 @@ class TestUnderscoreHandling(TestCase):
             catch_underscore_syntax("")
         except Exception as e:
             self.fail(e)
-
-
-class TestOldPapernumDetection(TestCase):
-    """Test :func:`.is_old_papernum`."""
-
-    def test_is_old_papernum(self):
-        """User enters a 7-digit number that looks like an old papernum."""
-        self.assertFalse(is_old_papernum('9106001'))
-        self.assertTrue(is_old_papernum('9107001'))
-        self.assertFalse(is_old_papernum('9200001'))
-        self.assertTrue(is_old_papernum('9201001'))
-        self.assertTrue(is_old_papernum('0703999'))
-        self.assertFalse(is_old_papernum('0704001'))
