+++ search/services/index/tests/tests.py
@@ -8,7 +8,7 @@ from elasticsearch_dsl.query import Range, Match, Bool, Nested
 
 from search.services import index
 from search.services.index import advanced
-from search.services.index.util import wildcardEscape, Q_
+from search.services.index.util import wildcard_escape, Q_
 from search.domain import Query, FieldedSearchTerm, DateRange, Classification,\
     AdvancedQuery, FieldedSearchList, ClassificationList, SimpleQuery, \
     DocumentSet
@@ -125,7 +125,7 @@ class TestWildcardSearch(TestCase):
     def test_match_any_wildcard_is_present(self):
         """A * wildcard is present in the query."""
         qs = "Foo t*"
-        qs_escaped, wildcard = wildcardEscape(qs)
+        qs_escaped, wildcard = wildcard_escape(qs)
 
         self.assertTrue(wildcard, "Wildcard should be detected")
         self.assertEqual(qs, qs_escaped, "The querystring should be unchanged")
@@ -138,7 +138,7 @@ class TestWildcardSearch(TestCase):
     def test_match_any_wildcard_in_literal(self):
         """A * wildcard is present in a string literal."""
         qs = '"Foo t*"'
-        qs_escaped, wildcard = wildcardEscape(qs)
+        qs_escaped, wildcard = wildcard_escape(qs)
 
         self.assertEqual(qs_escaped, '"Foo t\*"', "Wildcard should be escaped")
         self.assertFalse(wildcard, "Wildcard should not be detected")
@@ -151,7 +151,7 @@ class TestWildcardSearch(TestCase):
     def test_multiple_match_any_wildcard_in_literal(self):
         """Multiple * wildcards are present in a string literal."""
         qs = '"Fo*o t*"'
-        qs_escaped, wildcard = wildcardEscape(qs)
+        qs_escaped, wildcard = wildcard_escape(qs)
 
         self.assertEqual(qs_escaped, '"Fo\*o t\*"',
                          "Both wildcards should be escaped")
@@ -165,7 +165,7 @@ class TestWildcardSearch(TestCase):
     def test_mixed_wildcards_in_literal(self):
         """Both * and ? characters are present in a string literal."""
         qs = '"Fo? t*"'
-        qs_escaped, wildcard = wildcardEscape(qs)
+        qs_escaped, wildcard = wildcard_escape(qs)
 
         self.assertEqual(qs_escaped, '"Fo\? t\*"',
                          "Both wildcards should be escaped")
@@ -179,7 +179,7 @@ class TestWildcardSearch(TestCase):
     def test_wildcards_both_inside_and_outside_literal(self):
         """Wildcard characters are present both inside and outside literal."""
         qs = '"Fo? t*" said the *'
-        qs_escaped, wildcard = wildcardEscape(qs)
+        qs_escaped, wildcard = wildcard_escape(qs)
 
         self.assertEqual(qs_escaped, '"Fo\? t\*" said the *',
                          "Wildcards in literal should be escaped")
@@ -193,7 +193,7 @@ class TestWildcardSearch(TestCase):
     def test_wildcards_inside_outside_multiple_literals(self):
         """Wildcard chars are everywhere, and there are multiple literals."""
         qs = '"Fo?" s* "yes*" o?'
-        qs_escaped, wildcard = wildcardEscape(qs)
+        qs_escaped, wildcard = wildcard_escape(qs)
 
         self.assertEqual(qs_escaped, '"Fo\?" s* "yes\*" o?',
                          "Wildcards in literal should be escaped")
@@ -208,7 +208,7 @@ class TestWildcardSearch(TestCase):
     def test_wildcard_at_opening_of_string(self):
         """A wildcard character is the first character in the querystring."""
         with self.assertRaises(index.QueryError):
-            wildcardEscape("*nope")
+            wildcard_escape("*nope")
 
         with self.assertRaises(index.QueryError):
             Q_('match', 'title', '*nope')
